Auth: Do not stop processes that were never started
It blocks the process unnecessarily

+++ src/auth/Auth.cpp
@@ -383,6 +383,10 @@ namespace SDDM {
     }
 
     void Auth::stop() {
+        if (d->child->state() == QProcess::NotRunning) {
+            return;
+        }
+
         d->child->terminate();
 
         // wait for finished
